+++ plugins/node/opentelemetry-instrumentation-pino/test/pino-enabled.test.ts
@@ -18,17 +18,30 @@ import * as assert from 'assert';
 import * as semver from 'semver';
 import * as sinon from 'sinon';
 
-import { INVALID_SPAN_CONTEXT, Span, context, trace } from '@opentelemetry/api';
+import { INVALID_SPAN_CONTEXT, context, trace } from '@opentelemetry/api';
 import {
   InMemorySpanExporter,
   SimpleSpanProcessor,
 } from '@opentelemetry/sdk-trace-base';
-import { PinoInstrumentation, PinoInstrumentationConfig } from '../src';
+import {
+  assertInjection,
+  assertRecord,
+  init,
+  instrumentation,
+  kMessage,
+  logger,
+  pino,
+  setLogger,
+  setup,
+  stream,
+  testInjection,
+  testNoInjection,
+  writeSpy,
+} from './common';
 
 import { AsyncHooksContextManager } from '@opentelemetry/context-async-hooks';
 import { NodeTracerProvider } from '@opentelemetry/sdk-trace-node';
 import type { pino as Pino } from 'pino';
-import { Writable } from 'stream';
 
 const memoryExporter = new InMemorySpanExporter();
 const provider = new NodeTracerProvider();
@@ -36,79 +49,7 @@ const tracer = provider.getTracer('default');
 provider.addSpanProcessor(new SimpleSpanProcessor(memoryExporter));
 context.setGlobalContextManager(new AsyncHooksContextManager());
 
-const kMessage = 'log-message';
-
 describe('PinoInstrumentation', () => {
-  let stream: Writable;
-  let writeSpy: sinon.SinonSpy;
-  let pino: typeof Pino;
-  let instrumentation: PinoInstrumentation;
-  let logger: Pino.Logger;
-
-  function assertRecord(
-    record: any,
-    span: Span,
-    expectedKeys?: PinoInstrumentationConfig['logKeys']
-  ) {
-    const { traceId, spanId, traceFlags } = span.spanContext();
-    assert.strictEqual(record[expectedKeys?.traceId ?? 'trace_id'], traceId);
-    assert.strictEqual(record[expectedKeys?.spanId ?? 'span_id'], spanId);
-    assert.strictEqual(
-      record[expectedKeys?.traceFlags ?? 'trace_flags'],
-      `0${traceFlags.toString(16)}`
-    );
-    assert.strictEqual(kMessage, record['msg']);
-  }
-
-  function assertInjection(
-    span: Span,
-    expectedKeys?: PinoInstrumentationConfig['logKeys']
-  ) {
-    sinon.assert.calledOnce(writeSpy);
-    const record = JSON.parse(writeSpy.firstCall.args[0].toString());
-    assertRecord(record, span, expectedKeys);
-    return record;
-  }
-
-  function testInjection(
-    span: Span,
-    expectedKeys?: PinoInstrumentationConfig['logKeys']
-  ) {
-    logger.info(kMessage);
-    return assertInjection(span, expectedKeys);
-  }
-
-  function testNoInjection() {
-    logger.info(kMessage);
-    sinon.assert.calledOnce(writeSpy);
-    const record = JSON.parse(writeSpy.firstCall.args[0].toString());
-    assert.strictEqual(record['trace_id'], undefined);
-    assert.strictEqual(record['span_id'], undefined);
-    assert.strictEqual(record['trace_flags'], undefined);
-    assert.strictEqual(kMessage, record['msg']);
-    return record;
-  }
-
-  function init(importType: 'global' | 'default' | 'pino' = 'global') {
-    stream = new Writable();
-    stream._write = () => {};
-    writeSpy = sinon.spy(stream, 'write');
-    if (importType === 'global') {
-      logger = pino(stream);
-    } else {
-      // @ts-expect-error the same function reexported
-      logger = pino[importType](stream);
-    }
-  }
-
-  function setup(config?: PinoInstrumentationConfig) {
-    instrumentation = new PinoInstrumentation(config);
-    if (config?.enabled !== false) {
-      instrumentation.enable();
-    }
-    pino = require('pino');
-  }
-
   describe('enabled instrumentation', () => {
     beforeEach(() => {
       setup();
@@ -229,7 +170,7 @@ describe('PinoInstrumentation', () => {
     });
 
     it('does not fail when constructing logger without arguments', () => {
-      logger = pino();
+      setLogger(pino());
       const span = tracer.startSpan('abc');
       context.with(trace.setSpan(context.active(), span), () => {
         logger.info(kMessage);
@@ -239,7 +180,7 @@ describe('PinoInstrumentation', () => {
     });
 
     it('preserves user options and adds a mixin', () => {
-      logger = pino({ name: 'LogLog' }, stream);
+      setLogger(pino({ name: 'LogLog' }, stream));
 
       const span = tracer.startSpan('abc');
       context.with(trace.setSpan(context.active(), span), () => {
@@ -250,7 +191,7 @@ describe('PinoInstrumentation', () => {
 
     describe('binary arguments', () => {
       it('is possible to construct logger with undefined options', () => {
-        logger = pino(undefined as unknown as Pino.LoggerOptions, stream);
+        setLogger(pino(undefined as unknown as Pino.LoggerOptions, stream));
         const span = tracer.startSpan('abc');
         context.with(trace.setSpan(context.active(), span), () => {
           testInjection(span);
@@ -258,12 +199,14 @@ describe('PinoInstrumentation', () => {
       });
 
       it('preserves user mixins', () => {
-        logger = pino(
-          {
-            name: 'LogLog',
-            mixin: () => ({ a: 2, b: 'bar' }),
-          },
-          stream
+        setLogger(
+          pino(
+            {
+              name: 'LogLog',
+              mixin: () => ({ a: 2, b: 'bar' }),
+            },
+            stream
+          )
         );
 
         const span = tracer.startSpan('abc');
@@ -276,13 +219,15 @@ describe('PinoInstrumentation', () => {
       });
 
       it('ensures user mixin values take precedence', () => {
-        logger = pino(
-          {
-            mixin() {
-              return { trace_id: '123' };
+        setLogger(
+          pino(
+            {
+              mixin() {
+                return { trace_id: '123' };
+              },
             },
-          },
-          stream
+            stream
+          )
         );
 
         const span = tracer.startSpan('abc');
@@ -292,47 +237,6 @@ describe('PinoInstrumentation', () => {
 
         const record = JSON.parse(writeSpy.firstCall.args[0].toString());
         assert.strictEqual(record['trace_id'], '123');
-      });
-    });
-  });
-
-  describe('disabled instrumentation', () => {
-    beforeEach(() => {
-      setup({ enabled: false });
-      instrumentation.disable();
-      init();
-    });
-
-    // afterEach(() => {
-    //   instrumentation.enable();
-    // });
-
-    it('does not inject span context', () => {
-      const span = tracer.startSpan('abc');
-      context.with(trace.setSpan(context.active(), span), () => {
-        testNoInjection();
-      });
-    });
-
-    it('does not call log hook', () => {
-      const span = tracer.startSpan('abc');
-      instrumentation.setConfig({
-        enabled: false,
-        logHook: (_span, record) => {
-          record['resource.service.name'] = 'test-service';
-        },
-      });
-      context.with(trace.setSpan(context.active(), span), () => {
-        const record = testNoInjection();
-        assert.strictEqual(record['resource.service.name'], undefined);
-      });
-    });
-
-    it('injects span context once re-enabled', () => {
-      instrumentation.enable();
-      const span = tracer.startSpan('abc');
-      context.with(trace.setSpan(context.active(), span), () => {
-        testInjection(span);
       });
     });
   });
